requirements.txt
pytest
pytest-mock
flask
python-dotenv
